# src/frontend/src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Fitness Tracking API
  version: 1.0.0
  description: |
    A comprehensive fitness tracking API with voice integration, workout management, 
    exercise libraries, and LLM-powered workout generation capabilities.
  
    ## Features
    - User authentication with JWT tokens
    - Workout creation and tracking
    - Exercise library with media support
    - Voice-powered workout logging
    - LLM workout generation
    - Goal setting and progress tracking
    - Offline support with sync capabilities
    
  contact:
    name: API Support
    email: support@fitnessapi.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.fitnessapp.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/register:
    post:
      tags:
        - Authentication
      summary: User registration
      description: Create a new user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /auth/reset-password:
    post:
      tags:
        - Authentication
      summary: Reset password
      description: Reset password using reset token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # Workout Endpoints
...
  /exercises:
    get:
      tags:
        - Exercises
      summary: Get exercises
      description: Retrieve exercises with filtering and pagination
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: category
          in: query
          schema:
            $ref: '#/components/schemas/ExerciseCategory'
        - name: muscle
          in: query
          schema:
            $ref: '#/components/schemas/MuscleGroup'
        - name: equipment
          in: query
          schema:
            $ref: '#/components/schemas/Equipment'
        - name: search
          in: query
          schema:
            type: string
          description: Search term for exercise name or description
      responses:
        '200':
          description: Exercises retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PaginatedExercises'

    post:
      tags:
        - Exercises
      summary: Create exercise
      description: Create a new exercise (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExerciseCreateRequest'
      responses:
        '201':
          description: Exercise created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Exercise'
  /exercises/{id}:
    get:
      tags:
        - Exercises
      summary: Get exercise by ID
      description: Retrieve a specific exercise by its ID
      parameters:
        - $ref: '#/components/parameters/ExerciseId'
      responses:
        '200':
          description: Exercise retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Exercise'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  parameters:
...
    ExerciseId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Exercise ID
      example: "exercise_456"
...
  schemas:
...
    User:
      type: object
      required:
        - id
        - email
        - name
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          example: "user_123"
        email:
          type: string
          format: email
          example: "john@example.com"
        name:
          type: string
          example: "John Doe"
        avatar:
          type: string
          format: uri
          example: "https://example.com/avatar.jpg"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
    Exercise:
      type: object
      required:
        - id
        - name
        - category
        - movementPattern
        - primaryMuscles
        - secondaryMuscles
        - equipment
      properties:
        id:
          type: string
          example: "exercise_bench_press"
        name:
          type: string
          example: "Bench Press"
        description:
          type: string
          example: "A fundamental upper body strength exercise"
        category:
          $ref: '#/components/schemas/ExerciseCategory'
        movementPattern:
          $ref: '#/components/schemas/MovementPattern'
        primaryMuscles:
          type: array
          items:
            $ref: '#/components/schemas/MuscleGroup'
          example: ["chest", "triceps"]
        secondaryMuscles:
          type: array
          items:
            $ref: '#/components/schemas/MuscleGroup'
          example: ["shoulders"]
        equipment:
          type: array
          items:
            $ref: '#/components/schemas/Equipment'
          example: ["barbell"]
        instructions:
          type: array
          items:
            type: string
          example: ["Lie on bench", "Lower bar to chest", "Press up"]
        videoUrl:
          type: string
          format: uri
          example: "https://example.com/video.mp4"
        imageUrl:
          type: string
          format: uri
          example: "https://example.com/image.jpg"
...
    ExerciseCategory:
      type: string
      enum:
        - strength
        - cardio
        - flexibility
        - balance
        - mobility
        - plyometric
      example: strength
    MovementPattern:
      type: string
      enum:
        - squat
        - hinge
        - lunge
        - push
        - pull
        - carry
        - twist
        - gait
      example: push
    MuscleGroup:
      type: string
      enum:
        - chest
        - back
        - shoulders
        - biceps
        - triceps
        - forearms
        - core
        - glutes
        - quadriceps
        - hamstrings
        - calves
        - traps
        - lats
      example: chest
    Equipment:
      type: string
      enum:
        - barbell
        - dumbbell
        - kettlebell
        - cable
        - machine
        - bodyweight
        - resistance_band
        - medicine_ball
        - suspension_trainer
        - none
      example: barbell
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "john@example.com"
        password:
          type: string
          format: password
          example: "securePassword123"
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
        - confirmPassword
      properties:
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        password:
          type: string
          format: password
          example: "securePassword123"
        confirmPassword:
          type: string
          format: password
          example: "securePassword123"
...
    ResetPasswordRequest:
      type: object
      required:
        - token
        - newPassword
      properties:
        token:
          type: string
          example: "reset_token_123"
        newPassword:
          type: string
          format: password
...
    ExerciseCreateRequest:
      type: object
      required:
        - name
        - category
        - movementPattern
        - primaryMuscles
        - equipment
      properties:
        name:
          type: string
          example: "Bench Press"
        description:
          type: string
        category:
          $ref: '#/components/schemas/ExerciseCategory'
        movementPattern:
          $ref: '#/components/schemas/MovementPattern'
        primaryMuscles:
          type: array
          items:
            $ref: '#/components/schemas/MuscleGroup'
        secondaryMuscles:
          type: array
          items:
            $ref: '#/components/schemas/MuscleGroup'
        equipment:
          type: array
          items:
            $ref: '#/components/schemas/Equipment'
        instructions:
          type: array
          items:
            type: string
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              nullable: true
    AuthResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              required:
                - user
                - token
              properties:
                user:
                  $ref: '#/components/schemas/User'
                token:
                  type: string
                  example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    PaginatedResponse:
      type: object
      required:
        - data
        - total
        - page
        - limit
        - hasNext
        - hasPrev
      properties:
        data:
          type: array
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: false
...
    PaginatedExercises:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Exercise'
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Invalid input data"
            details:
              type: object
              example: {"field": "email", "issue": "Invalid email format"}
  responses:
...
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "UNAUTHORIZED"
              message: "Invalid or expired token"
...
    ConflictError:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "CONFLICT"
              message: "Resource already exists"
    ValidationError:
      description: Validation Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "VALIDATION_ERROR"
              message: "Invalid input data"
              details:
                email: "Invalid email format"
                password: "Password must be at least 8 characters"
...